requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tokenizers>=0.13.0
numpy>=1.24.0
pillow>=9.5.0
pyyaml>=6.0
tqdm>=4.65.0
einops>=0.6.0
accelerate>=0.20.0
wandb>=0.15.0
huggingface_hub
datasets
hf_transfer
